Experiments/requirements.txt
numpy==1.25.2
pandas==2.0.3
pandas_stubs==1.2.0.57
Pillow==10.1.0
PyYAML==6.0.1
salesforce_lavis==1.0.2
scipy==1.11.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transformers==4.25.0
